Add configs for the GitHub App
These configs are optional to make it easier to boot up the app even if
it is not fully configured.

The enabled flag allows us to toggle whether the GitHub app features are
enabled or not. If not all GitHub app configurations are set, this
toggle is automatically turned off.