Reorganize app, handlers and models
The purpose of this change is to associate Pydantic model modules with
an app (router) rather than be global across the APP. These individual
apps can implement different API versions, and it makes a great deal of
sense to co-locate those schemas with the associated app.

There was also a bug with having multiple apps. What we actually want is
a single app, but multiple routers. The external router is configured
with the app name path prefix.