Create MessageIdProtocol
The MessageIdProtocol (and BroadcastMessage.identifier attribute)
replace the previous source_path attribute. The goal is to abstract away
the identifier to be anything that's hashable, and with the Protocol
we've given it a semantic type.

This will allow multiple backends to introduce their own system for
uniquely identifying a message in a repository, without us needing to
depend on that detail within the domain model.