Use new logic for triggering CI
Only trigger CI on pull requests, pushes of tags, and pushes of
unexpected branches to allow a more typical CI pattern of running
CI on each pull request.  Add a comment to the logic for when to
do the Docker build from CI, and update that logic and the script
to generate the Docker tag to allow for pull requests.