Add BroadcastCategory to messages
The category is an "enum" that defaults to "maintenance" category, which
is our current usage. The new category that's now available is "info."
Apps like squareone can display messages from categories differently;
like red for a maintenance message, or Rubin primary teal for an info
message.

In markdown front-matter, this category can be set via the category key:

```
---
category: info
---
```

This is propagated into BroadcastMessage's category attribute. Finally,
the model for broadcast messages in the v1 API also propagates this
information in the "category" field.

For messages that don't set a category, the default is always
"maintenance", which is our current idiomatic treatment of messages.